fix(mock): 表格数据接口三合一
