+++ src/types/api/dictTree.d.ts
@@ -1,0 +1,8 @@
+declare namespace IDict {
+  interface ITbParams {
+    page: number;
+    size: number;
+    id?: string;
+    searchKey?: string;
+  }
+}
